requirements.txt
fastapi
uvicorn
requests
PyMuPDF
faiss-cpu
sentence-splitter
python-multipart
python-dotenv